refactor: refactors Router class and adds Routing
